[sherlock][audio] Add audio output support for EVT boards
EVT changes the output codecs compared to P2 from one TAS5760 for tweeters
plus one TAS5720 for the woofer, into 3 TAS5720s two for the tweeters and one
for the woorfer.

Test: Play test tone in a Sherlock P2 and in an EVT, and also
manually turn on one speaker at the time.
Note that the -c 4 parameter is required because we expect 4 channels input
L+R for tweeters + L+R for the 1 Woofer (mixed in HW).
$ audio gain -50
audio gain -50
Gain is now -50.00 dB.  Stream is unmuted.
$ audio -c 4 tone 440 .1
audio -c 4 tone 440 .1
Playing 440.00 Hz tone for 0.10 seconds

Change-Id: I81dd9f2edb150b55b94125755381054f3b7105ce