[guest] Fix vsock test connection shutdown race
When the vsock integration test shuts down and quickly opens a new connection it
will reuse the source port. This can lead to the vsock service in the guest becoming
confused as it may see the new connection request before it has completed cleaning up
the previous connection and reject the new connection as a spurious attempt to recreate
an existing connection.

This race occurs as the vsock test is shutting down its connections ungracefully and so
is unable to wait for the guest to confirm it has completed shutdown. The solution used
here is to ensure each connection is unique by using a unique listener, and hence listen
port, for each connection.

Test: run-test guest_integration_tests

Change-Id: Ie3cea59815a6a05bf4f2237f3012d6efcfc92d4e