[debugger] Simplify physical frame finishing.
Previously the finish controller could handle the case where the
fingerprint of the frame to finish was not synchronously available. This
complicated the initialization logic.

This change removes this capability. This means that only the top frame
(or one of its inlines) can be finished, or the stack must be fully
available.

Changes the console frontend to require the full stack to be available
when dealing with anything but the top physical frame. In the future the
frontend should be able to sync the full stack automatically when
needed. In the meantime this isn't a big limitation because the user
won't know which frame to select without printing the backtrace (which
will sync the stack).

TEST=manual,unit

Change-Id: I1a65017d0484927983257d3606d7d704e7d63ffd