[ledger][coroutine] fix use-after-free in debug check
CoroutineManager::StartCoroutine checked that the handler was removed
from its list to confirm that the callback passed to the coroutine has
been called. This does not work if the manager was destroyed by the
callback. This changes the check to use a stack-allocated variable.

TEST=test CoroutineManager.DeleteInCallback does not crash in asan,
StartCoroutine dchecks if the callback is not called

LE-673 #done

Change-Id: Iaaf96b63509e691b427fdd19c4ad4d508925cc97