[libc][statfs] Correctly stub out statfs, fstatfs, statvfs, and fstatvfs
statfs and fstatfs were already weak symbols. This change places the
weak libc stubs in the standard place.

In statvfs and fstatvfs, every part of their implementation was dead
besides the call to statfs/fstatfs. These are currently unimplemented
by fdio. The translation between the fields should be handled by fdio,
not libc, if reintroduced when implementing statvfs and fstatvfs.

Test: no functional change
Change-Id: Iaa8cfaa8f56777c8cd104be02f89c99cd8e8229f