[debugger] Set inline frames when stepping.
Adds support for setting a preferred inline frame in cases where they
are ambiguous (the instruction pointer is at the beginning of an inline
function). A function for setting this is added on the ThreadController,
and a unit test is added for this functionality.

The new function is used by the StepOver and Finish controllers and
tests are added for these case.

Move the helper SetAddressForMockFrame to MockFrame.SetAddress().

TEST=unit,manual

Change-Id: Id399d5d15ef5193ea254ef3aaa24af05dff628fd