[dev][acpi] (First step in) Move ACPI under platform bus (Abort, Retry...)
This is the first step in instituting a x86 platform bus board driver,
by moving the boot ACPI code (namespace walk, initializing ACPI drivers)
and the ACPI drivers under the platform bus. The ACPI boot code (+ drivers)
have been converted to a platform bus board driver, which binds based on
the PDEV/VDEV, cons'ed up early in boot. The next phase is to convert
each of the ACPI drivers to a platform bus board driver (still running
in the same sys devhost). The final phase is to identify which of the
ACPI drivers make sense to move to (PCI ?) it's own devhost, with all
of the other ACPI devices in yet another devhost, and implement the bus
protocol to proxy ACPI related calls across the devhosts. The pci node
is created under /dev/sys/ to preserve compatibility, and acpi/ is
created under /dev/sys/platform.

Bug: ZX-2655

Test: CQ. Boot and make sure that all ACPI devices are initialized. lspower
works, verify dm dump shows the devices in the correct hierarchy.

Change-Id: Id868e8a6efc1fccb3d89a9a001fcf117ffd8315f