[bt][l2cap] PDU: allow Copy()-ing zero bytes
As part of the upcoming L2CAP channel modes support, we'll need to
support segmentation and reassembly. And to support segmentation and
reassembly, we need to make PDUs and SDUs distinct entities.

After some discussion (see BT-61), we decided to represent SDUs as
simple ByteBuffers, rather than something more complicated
(e.g. linked-list of PDUs). This means that the Channel object will
soon provide a ByteBuffer when it calls |rx_callback|, rather than a
PDU.

Since various existing |rx_callback|s expect to be able to receive a
PDU with an empty body, we need to support passing an empty L2CAP
frame body up as a ByteBuffer. To that end, remove the assertion in
PDU::Copy(), which requires that the PDU have non-zero bytes available
past |pos|.

Bug: BT-716
Test: fx run-test bluetooth-tests -t bt-host-unittests

Change-Id: I62f467b4c39f46e31b97475ed5a3910ce28ba887