[zircon][fvm] Decouple format from disk size.
This is the first step towards allowing gaps between primary
and secondary metadata. This gaps are meant to represent room
for growth when the underlying block device grows (extend the fvm
partition).

added fvm/test for unittests.

The integration test that was added verifies that even if the underlying
block devices grows, for now, the FVM remains unchanged in terms of
addressable slices and data persistency. It is just a safety check as we
add logic to actually grow and increase the addressable slices later.

TEST=fvm-unit-test, fvm-test

Change-Id: I733aea9f22431d98dfaa792e52ebe80a2d9cf59e