[guest] Explicit dispatcher usage
This makes the choice of dispatcher for bell traps and FIDL messages more explicit.
As part of making the choice explicit the 'device_dispatcher' is removed from Guest
and instead a dispatcher must be passed to all locations that would previously reach
for the `guest->device_dispatcher`.

After this change all out of process device messaging will go through a dedicated
`device_loop` and other IPC for external processes goes through the main `loop`.
The two temporary exceptions are virtio-net-legacy, which uses the `device_loop`,
and the `virtio-vsock`, which uses its own dedicated dispatcher.

Test: run-test guest_integration_tests

Change-Id: I01dd1d878d9db6042e4c2774bec41d5129899780