[bt][l2cap] replace references to SDU with ByteBufferPtr
Now that we've updated client code to receive l2cap::Channel
data with a ByteBuffer instead of a PDU, remove the SDU
typedef.

Also: when invoking a callback with the content of a
service-data-unit, move ownership of the data (which is now in a
ByteBuffer, rather than an SDU-typed-object) to the callee.

Bug: BT-716
Test: fx run-test bluetooth-tests -t bt-host-unittests

Change-Id: I8e889b88bf00d1f78d6de14ba9a14adbad8d591a