[media][tests] Test cleanup and clang-tidy
This CL addresses a small number of clang-tidy issues, renames a
header, cleans up a few compile-time dependencies, and clarifies
some sections that are prone to race conditions if misused -- all
toward the goal of preping the codebase for subsequent CL that adds
a new test binary to validate fuchsia.media.AudioDeviceEnumerator.

It also doubles the timeout interval used in numerous checks, when
there is no FIDL response callback, and the Disconnect event would
be considered a test case failure.

Test: build, CQ, git-file-tidy

Change-Id: I2b6e6bea459e62ee04b1399271d41c4c3d8bf7aa