[audio][tests] AudioDeviceEnumerator tests
This CL adds the initial set of test cases for the FIDL interface
fuchsia.media.AudioDeviceEnumerator. These 15 test cases work
entirely non-intrusively with whatever existing devices are found
on the device-under-test, and for that reason cover only a
minority portion of the entire interface. However, they form a
starting point from which future virtual-device-based tests can be
created. These tests are found in the newly-added test binary
'audio_device_tests'.

test: build, CQ (which includes 'audio_device_tests')

Change-Id: Ie18d85210b5836d9d9164ce68a4508ba29d291bb