[fidlc] Include source location in IR
This change introduces a location field in the JSON IR in each const,
enum, interface, struct, union, table and xunion, their members and
their arguments indicating where that item is declared. It includes the
filename (as supplied to fidlc), line number and column number.

This allows tools that operate on the IR such as bindings generators and
analysis tools to refer to where declarations were initially made,
typically for error reporting.

TEST=out/build-zircon/host_tests/fidl-compiler-test; fx build

Change-Id: I60badc276c5746eefd0dd96b483ed552b54cb0ef