[docs][fidl] Add prototypes to C bindings documentation.
I was using these APIs, and I found it time-consuming to have to
flip back and forth between the documentation and the method
signature to figure out what the various identifiers were.

Another approach would be to remove the docs from this .md file and
put them in the relevant .h file.

Test: Looked at gitiles.
Change-Id: I914c210e8b297d27cb666d757961cd2413994f5c