[escher] Final prep for Scenic/PaperRenderer2 integration.
The biggest change is for PaperRenderer2 to always use the
"latched pose buffer" representation for camera matrices,
and to use a Vulkan push constant to choose the left or
right eye matrix.  This is the easiest way to maintain
compatibility with the old PaperRenderer.

Escher's CommandBuffer abstraction has a limitation where
it stomps on binding-resource IDs when beginning a render
pass.  This CL adds DCHECKs to detect such misuse.

Various Scenic clients (root_presenter and examples) are
updated in anticipation of the PaperRenderer2 integration.
For example, hello_scenic takes a command-line argument
to specify the shadow-type to use, and also exercises clipping
using the recently-added SetClipPlanesCmd.  Also, the upcoming
"stencil shadows" will initially only work with PointLights,
so clients add these in addition to the existing ambient
and directional lights.

SCN-670 #comment

TEST=
hello_scenic, hello_base_view, hello_pose_buffer, hello_stereo
escher_unittests (host and target)

Change-Id: I3e9cc9576e1231321044df7af84a8a3f61e86980