[Terminal] Parse escape sequences properly
The terminal was previously using the wrong part of the API on the
term_model to insert text, specifically term_model::term::Term.input,
which meant that the characters fed in were just echoed to the character
grid, without parsing any escape sequences. This fixes that by using the
term_model::ansi::Processor class, which parses the escape sequences and
then delegates to term_model::term::Term to do the actual work.

Testing: Manual. Launch Ermine, login as guest, hit capslock+space to
bring up the Ask bar, type "terminal" to open the terminal. Observe that
some of the test string is colored because escape sequences are parsed.
Also characters typed on the keyboard are still echoed to the screen.

Change-Id: Id58abfa5a1add8c85f38b8dd365d6e7c8a74cf27