[auth] Add a human readable name to the account API.
<This change was previously created and approved in garnet as
cl/247019 prior to the repo refactoring. The only modification is I've
used _SIZE rather then _LENGTH per the new API council preference>

This will be important for AccountManager to replace the functionality
of the UserProvider. Eventually its likely we'll need to also add an
account picture, but lets delay that until there is a consumer.

The current implementation returns a fixed string. Once we have
persistence and accounts hooked up we should populate from the service
provider account during ProvisionFromSericeProviderAccount. Long term we
should probably integrate with a ledger page to synchronize name between
devices and this need for synchoronization is why I'm currently only
adding
a getter not a setter.

Tested: CQ
Change-Id: I5829dd3d2a20b4aa75df8d870a2bb396d946f90d