[appmgr] Stop providing minfs root to components requesting persistent storage
This change adds a new name for the existing 'persistent-storage'
feature, 'deprecated-global-persistent-storage', which existing
components will be migrated to before deleting the old
'persistent-storage' name.

Along with that, it modifies the behavior of these features to provide a
subdirectory of minfs instead of the root.  All components that request
these features will receive the same non-isolated "global" directory as
"/data" in their namespace. This prevents components using the
deprecated feature from accessing arbitrary contents in minfs, such as
the storage of other components using the new
'isolated-persistent-storage' feature.

So that this change is OTA-safe, existing data in the root of minfs is
moved into the new subdirectory with the exception of specific
files/directories that are currently stored in minfs and accessed
outside of the persistent storage features (specifically, the SSH
authorized_keys and the dynamic pkgfs index).

CF-28 #progress

Test: Paved system with build made before this change, applied this
change, built, and ran 'fx ota'. Verified that device rebooted and data
was migrated correctly without losing ssh access and without migrating
the pkgfs_index.

Change-Id: I65552b8b5d68b38e0fbcaaff1b7e1e222b9e7c5b