[debugger] Improve resume count checking in tests.
Previously tests would check the total count of resumes sent to the
agent as a way to test whether the program has resumed or not. This
stateful method made it diffult to follow what the value is supposed to
be.

Changes the getter for the resume count to reset the value. The test
code can now check for 1 (resumed) or 0 (stayed stopped) which makes the
test easier to follow and modify.

TEST=this is

Change-Id: I71654b2eaa549b6780eef6cd76f0f65a2cc320ca