[kernel][lib] Add rand_r
Add a version of the rand-like API which allows users to provide their
own seed/core-state.  This is only really to support testing scenarios
where we just need "OK" pseudo-randomness, but need to be able to
control the seed/core, instead of sharing global state.

Tests: Build only for this CL.  Exercised in subsequent CLs.
Change-Id: Ie161970d10ba64ea937a09802e59ae730b5ee1d5