[perfcompare] Add new entry point for perfcompare recipe to invoke (reland)
Add benchmarks_perfcompare.sh.

This initial version runs a subset of the tests that benchmarks.sh
runs, because the full set takes too much time for the bot to run
twice without going over the bot's current timeout limit.  Running a
faster set of tests is also easier for development testing of the
perfcompare recipe.

This is a reland of the change.  When it was committed before, the
change was incorrectly blamed for a failure and got reverted.

Bug: IN-646
Test: tested via 'led' with recipe change
Change-Id: I3013be663087359e6fa6d8d5d7e25c50d73b6f50