[perfcompare] Add new entry point for perfcompare recipe to invoke
Add benchmarks_perfcompare.sh.

This initial version runs a subset of the tests that benchmarks.sh
runs, because the full set takes too much time for the bot to run
twice without going over the bot's current timeout limit.  Running a
faster set of tests is also easier for development testing of the
perfcompare recipe.

Bug: IN-646
Test: tested via 'led' with recipe change
Change-Id: I54349d462423cd89dc051f0ae69d91f908793842