[fidl] Update wire format for optional xunions.
The wire format for an optional xunion inside a struct was a pointer to
the xunion, which resulted in a double-indirection to the xunion data
(pointer to xunion, followed by a pointer to the data in the xunion's
envelope); it should be an inline xunion instead. This matches the Rust
implementation.

Test: # added typeshape test for xunions to max_bytes_tests.cpp
Test: make HOST_USE_ASAN=true tools && $FUCHSIA_DIR/zircon/build-x64/host_tests/fidl-compiler-test

Test: # added xunion wire format tests to roundtrip_test.cc
Test: fx build-zircon -A -H -l && fx build host_x64/fidl_cpp_host_unittests && $FUCHSIA_DIR/out/x64/host_x64/fidl_cpp_host_unittests

Test: fx build garnet/go/src/fidl && fx exec $FUCHSIA_DIR/garnet/go/src/fidl/compiler/backend/typestest/regen.sh

Change-Id: If934dfb544479c1b9190abc024bcf5cc6e43a507