[debugger] Step into inline frames.
Implements "step into" for inline frames. This operation does not
actually step the CPU (since the frame is inlined there is no call), but
rather emulates stepping into a call by adjusting the visibile inline
frames in the stack.

A new controller is added to implement a "user level step into" which
calls through to the regular step controller for the normal case.

TEST=unit

Change-Id: Ia1b6c57ff86e3a7be1707a0b792f112c1697f263