[System Monitor] Fix samples used and crash
This CL changes which samples are used in a given GUI column. The
prior math was inclusive on both ends; the new math is exclusive on
the later value. This means that samples that were on the boundary
between two columns where counted in both columns. Now they are no
longer being double counted.

Second, this CL checks against failing to find the *any* samples in
a range, which could lead to a crash.

Test: garnet/lib/system_monitor/dockyard/dockyard_unittests.cc
Change-Id: I8d7f686c4236e75e5e6628d9420148a4601d7fbd