[tracing] Make trace a component package
With the introduction of shell binaries, now we can upgrade the
trace package to be a component and remove the
|deprecated_system_image| flag.

This also allows us to make trace a package that can be pushed
independently on the target.

bug: BLD-315

Tested: Tested manually by repaving the target and making sure that
trace can be started from shell.
fx shell trace record # records to default/persistent location /data/trace.json
fx shell trace record --output-file=/tmp/trace.json # records to system-temp

This patch is based on
https://fuchsia-review.googlesource.com/c/garnet/+/224295

Change-Id: Ic8beb5900cc1d6b338c98e130f7e8aa6eaf4ed7e