[package][asan] Fix test spec dependency bug
This fixes a bug in which non-existent test spec targets found their way
into the build graph when building with asan. A variable `is_disabled`
was though to be updated, which would later preclude the target from
entering the graph, when in actuality it was only a new, shadowed
`is_disabled` that was set.

Test: gn gen with asan does not result in conflict - and the correct
test spec metadata is still produced.

Change-Id: I5e33505e2baad607d545bb9e842987e658ed8339