[appmgr] packaged components need a package loader
There was an attempt to make programs that have "features":["shell"] able to
spawn subprograms by providing the system loader to them, but this is a
misfeature. It might make a handful of cases work today, but they don't work
tomorrow. Most critically it blocks asan deployment, because the build system
is unaware of this behavior and does not put the asan libraries of the
package into the system image package. As the system image package is going
away, this behavior really needs to be removed.

Bug: CF-576 #done
Test: CQ + topaz tryjobs
Test: lsblk and other commands work under SSH
Test: lsblk and other commands work in qemu serial console
Test: lsblk and other commands work in run-vc
Change-Id: I2677630334a05e60216a2641802190cd5025c3b5