[rust] Helpful error for disabled dependency
When a dependency is conditionally disabled in rustc_deps/Cargo.toml,
current script used to just fail in a key lookup.  The error message is
quite confusing.

TODO: It would be nice to show gn targets, instead of the JSON data file
paths in the error message, but I can not figure out how to do that.  Do
we need to pass those to the script as additional arguments?

Test: I was trying to use a dependency disabled for Fuchsia build
target.  With this change I see this new error message.

Change-Id: I8ae926e5b7bee2b486b0a370336a31edc1f98de0