[gn][banjo] Fix some signedness comparison warnings in gn host build
This ends up being a lot of ugly casts, but attempting to change the
string indices to size_t was not clearly correct in all cases either. So
this is just getting rid of warnings while avoiding changing behaviour
in code the might be abandoned in the near future anyway.

ZX-3415 #comment [gn][banjo] Fix some signedness comparison warnings in gn host build

Test: CQ
Change-Id: Ie49031cb89ea36f9a76fdbc2d1a843209d0be8fc