[wlan][mlme] Move buffer.rs into wlan-mlme
buffer.rs was declared in wlan-mlme-c as there was no possibility
to generate bindings for a crate's dependencies. I finally figured
out that one has to explicitly declare usage of another crate for
cbindgen to parse the crate via `extern crate foo`. Declaring the
dependency in the crate's Cargo.toml is not sufficient.
With being able to generate bindings for wlan-mlme buffer.rs can
be moved into wlan-mlme and be used directly rather than depending
on a trait which introduces an additional layer of abstraction.

Corresponding cbindgen bug:
https://github.com/eqrion/cbindgen/issues/221

Test=CQ

Change-Id: I9113a1a9b5d2e0e684dada7de9e5310ff407aea6