[debugger] Handle inlines when finishing frames.
Rename the Stack's "top" inline frame hide count to be "ambiguous"
inline frame hide count. The previuos API let you hide inline frames the
code is currently in the middle of, when only ambiguous locations should
have the stack adjusted.

The checking function for ambiguous frame locations was moved to the
Frame object.

Some commented-out code is added to the physical frame finish
controller. This needs some significant enhancements to the inline test
data which will require updating the other tests that use it. This will
be done in a followup.

Sets the inline location when doing an "until" operation. Adds a new
mode when setting the inline location to set it to one before the given
fingerprint.

Change-Id: I24a712e0632abece59badb1880cf37969c213414