[audio_core] Clean up driver/device shutdown sequence
In certain abnormal-exit code paths, input drivers would not be
given the opportunity to correctly clean up.  Although RAII meant
that no resources would be leaked, nonetheless this could lead to
inconsistent or incorrect behavior in the future (even worse, it
would be happening specifically when trying to recover from the
unexpected problem that caused the device error in the first
place).  With this change, drivers/devices are correctly cleaned
up and shut down in all foreseen normal or error exit paths.

Conversely, shutdown paths for audio output devices contain a
double-Deactivate of its two dispatcher channels. Previously
this led on heavily-loaded debug builds to an overactive
ASSERT in a lower-layer component; this has since been fixed.

This CL moves the call to the driver resource cleanup code to
a single, more central spot where it is guaranteed to be called
whether the exit was originated from the device, the driver, or
the device manager, and whether this error occurred while in the
device's execution domain or not. With this change, Input and
Output sides mirror each other.

Test: build, CQ, overnight loop of the new test case in question
(OnDefaultDeviceChanged after RemoveDevice)

Change-Id: Iff2e52e0ccf7cbb4cc10c4855e364b33cf5f0c85