[Terminal] Add initial unit tests
This actually just plugs in the tests in term-model. The reasoning for
this is that unit tests test internal, hermetically isolatable units
within the program. They do not test connections to anything outside the
program. Pretty much everything in main.rs is focused on setting up and
managing connections to external channels. All the meaningful business
logic which can be tested with unit tests is in term-model.

Further, while term-model may ostensibly be "third party" code, because
it was extracted from Alacritty (with permission), it's effectively been
forked already. Alacritty didn't separate out this part of their code
into a library which could be easily reused. We did that, and the code
had to be tweaked to allow it to stand on it's own. And we will likely
fork it much further in the future, as any significant changes to the
underlying terminal setup will require modifying term-model. There will
likely come a point where it's sufficiently ours that we will want to
stop calling it third party. So I think running automated tests on this
"third party" code makes sense.

TEST = fx run-test terminal_tests

Change-Id: Ia6feffceb845790fd273adb3266653e33fd53550