[modular][cleanup] Remove peridot/lib/common/names.h
Most constants are only used in one place, so moved where they are used.
The exception is kPresentationService, which is used in two files in
`basemgr`, but also indicates it is no longer necessary and can be
removed. Both new references cross-reference each other and we can
expect them to go away, but this is kept as a separate change.

TEST=compiles

Change-Id: I480b19ee1bdd16d4656d1d2a94b547aff76bc804