[sherlock][audio] Add audio recording support for EVT boards
EVT changes the number of microphones compared to P2 by removing the eq mic.
In this change we reduce the number of channels specified from 4 to 2,
using only those that come from the L+R microphones.  This disables the eq
mic even for P2 boards since we don't intent to support this
configuration and not having this mic enabled does not break other usage
of P2 boards.  Note that the comment in the code mentioning 4 mics was
incorrect, we had 3 mics but where outputing 4 channels.

Test: record a wav file, check validity and listen to contents on P2 and
EVT boards.
$ audio record /tmp/noise.wav 10
audio record /tmp/noise.wav 10
Recording for 10.0 seconds

Change-Id: Ie5a28239c61b8631a488d95cb48c8cd541960c10