[dispatcher-pool] Fix the preconditions for an ASSERT.
Fix the conditions which can cause a EventSource to fire an ASSERT
inside of InternalDeactivateLocked.  Also, clean up a few stale
comments which had become out of date.

Background:

When an event source becomes deactivated, it is supposed to guarantee
that it never fires again, and releases all of its resources.  There
is an edge case which must be carefully handled, however.  If the user
attempt to deactivate the event source (either from within the
execution domain or not), and there had been a pending wait (the state
is WaitingOnPort), but the kernel has already released a thread from
the pool in order to service the handle, then the Deactivate code
cannot safely reclaim the reference which is implicitly being held by
the kernel as part of the wait operation.

Instead, we need to wait until the thread makes it out of the
port-wait operation, reclaims the reference, and realizes that the
event source has become deactivated (because the domain_ reference has
been released).  When this happens, the work will not be queued to the
pending work queue for the event source, and the thread will drop the
reference to the EventSource as it unwinds.  Once all of the
references to the EventSource have gone away (including user held
references), the object will destruct and all will be well.

The ASSERT:

Inside DeactivateInternalLocked, the code was attempting to make an
assertion about the state of the system.  In particular, if the
EventSource no longer had a domain (!is_active()), then the deactivate
operation should become a no-op.  In this no-op case, the code wants
to sanity check the state of the system, so it was trying to assert
that...

1) The handle was still valid, but the domain was not, and we were in
   the WaitingOnPort state.  IOW - we are waiting for that last
   callback from the kernel to show up so that we know that the kernel
   ref has been released.  Or...
2) The handle is no longer valid, and we are either in the Idle state,
   or the Dispatching state.  It is valid to deactivate an event
   source from within a Dispatch operation, but other properties of
   the system should ensure that we are not in the DispatchPending
   state (previous calls to Deactivate should have guaranteed this).

The Bug:

The problem happens in that small edge case when the kernel thread
finally pops up and discovers that we are no longer active because our
domain has gone away.  In this situation, it returns us to the Idle
state, but does not bother to close the handle, assuming that it will
be closed when the object finally destructs.

The trouble with this is that it leaves us in a state where our domain
is gone, and our state is Idle, but we still have a valid handle.
At this point in time, none of the dispatcher framework code is
holding any references to our event source (kernel has released their
implicit reference, we are no longer in the event source list for our
domain), but _users_ might still be holding a reference to a now dead
object.  Deactivate should continue to be idempotent, even in this
state, but calling it will trip the assert described above.

The Fix:

Simply close the handle in the case where we transition from
WaitingOnPort to Idle.  This restores the invariant we are attempting
to assert above.

Future Work:

In theory (according to comments), EventSources are _not_ supposed to
be re-activatable.  Once actived, they are not supposed to be
activatable again.  By clearing the handle out here, we are actually
returning the object to a state where it could be used again.  As it
turns out, however, this was happening in the common case already.

At some point, it would be good to clarify this stance.  There is
nothing technically preventing re-use of the event source after it had
been deactivated.  It might be best to simply remove the comment.

Alternatively, if this is a property which should be maintained, then
it should be enforced.  Either adding a flag to indicate that the
object had once been activated, or just adding a new state to the
existing set of states would do the job.  Mucking with the state
machine here, however, requires more careful analysis.  This approach
preserves the ASSERT sanity check, and actually brings the EventSource
into the state that the typical path already does, so it is the low
risk fix for now.

ZX-3424 #done
Tests: Isolated and verified cause with instrumentation.  Verified fix
       using repro steps listed in ZX-3424

Change-Id: Ic3449cf15e32d91d36a454fda6af42d0900ea035