[debian_guest] Configure Debian for autologin on serial
We would like to use Debian instead of Linux for the integration tests
(this will make things easier for the network team's testing) and the
tests are simpler if don't need to log in. Also stop printing motd and
the issue (which is not useful for autologin anyway).

MAC-245 #comment

TEST=Launch debian_guest on x64 and arm64

Change-Id: I70bc52744acfb421c978cd763f5fa0ed19a7a98e