[ledger] GetHeadCommitIds returns a list in timestamp order
This was already true for PageStorageImpl but not FakePageStorage. This
change defines the order storage::Commit::TimestampOrdered, that
compares the timestamp them the commit id, documents that
GetHeadCommitIds returns a list sorted according to TimestampOrdered,
and changes FakePageStorage to enforce this order. Assertions that
checked timestamp ordering have been converted to use TimestampOrdered
directly.

TEST=PageDbTest.OrderHeadCommitsByTimestamp{,ThenId}

Change-Id: Ie99fcc2338b166636349796dfa956697302a6a17