[iquery] Fix goldens test to work with ASAN.
Our loading of iquery and sh from the iquery goldens test is not
compatible with ASAN, which requires all binaries to be packaged
hermetically with the test package.

While iquery can be included this way, sh cannot because it is a system
uapp in Zircon.

This change fixes the problem by making the following changes:
- Package iquery with the goldens test.
- Add an option to iquery to first change directories to a specific
location (one of the uses of sh).
- Change the golden test to perform the ID substitution necessary for
the outputs to match the golden file (the other use of sh, which piped
output through sed).
- Change all golden file input strings to just provide arguments to
iquery, not any additional shell commands.

CF-586: #done

Change-Id: I209318e9ceed4afb6714b1f03a91447b4cd16a19