[gn][fidlc] Suppress sign-compare warnings in flat_ast.h
It's not clear (at least to me) if it's always safe to cast away the
sign-compare warnings. For now, punt on this and revisit once we have a
safemath library that can be used to check the conversions easily.

ZX-3415 #comment [gn][fidlc] Suppress sign-compare warnings in flat_ast.h
FIDL-487 #progress

Test: CQ
Change-Id: Ibf2748219a2cf8518f3aeef598c5a2ebaa7dd0f6