[mdns] fix bug in address fix-ups
This CL fixes a bug in which the address fix-up logic assumed that
std::remove_if moves 'removed' entries to the end of the vector. In
fact, std::remove_if removes those entries and leaves the entries
at the end of the vector in a 'moved from' indeterminate state.

TEST: use avahi to resolve addresses

NET-2082 #done

Change-Id: I5b0887ba38c55c4c785f3030b16e8f4a7edf778a