[debugger] Enhance inline function finishing.
Enhance the Stack to not add ambiguous inline calls except for the top
physical frame of the function.

The address of a frame when it's not on the top of the stack is the
return address which is after the call. This means that if an inline
function starts at that address (the ambiguous case), it wasn't involved
in the function call of the frame above it on the stack.

Update the test data to reflect the above requirements, and the tests
that relied on this data.

Collapse inline frames when doing "finish" on a physical function call.
This is for the same reason as above: when completing a function call
any inline function just starting were not involved in the call.

Change-Id: I277d2ce1bcc5f403587460802418b0e9c2f350c3