[arm64][HW breakpoints][debugger] Enabling HW breakpoints for zircon.
This CL holds the implementation of HW breakpoints for Zircon for arm64.
It keeps correct track of the debug state per thread and manages the
correct mdscr values both for context switching and IRQ handling.

This also includes an userspace program used by the debugger to verify
that this exception is easily triggered. This utility program will be
used for debugger development in the future.

TEST=Run "hw_breakpointer" is the shell. Tested on a VIM2.

Change-Id: I5d22aee7266b54130083e2f8698f1c22051e6736