[media] Clarify FIDL disconnect messages
Per MTWN-237, update the FIDL disconnect messages for the
primary audio utilities for developers (signal_generator,
vol, wav_recorder) and the Audio (policy) service, to make
it clear that when the AudioCore service disconnects, these
are failures of the client connection, which might simply
be a client permissions issue (as opposed to a deeper system
failure).

Also, for the three audio utilities, display the error code
that has been passed to the error handler as the "epitaph".

BUG: MTWN-237 #done

TEST: build, CQ, manually verify error messages from vol,
signal_generator and wav_recorder on qemu without -g (thus
running in the serial console, without access to AudioCore).

Change-Id: I2980be6c44b9624c3ee40954c881655fc8078eb8