[cmc] Fix and re-enable broken cmc tests
The tests were broken by the roll of the valico crate to 2.4. Likely
breaking change:

https://github.com/rustless/valico/commit/902c2f115768c321221fc1dfabf9b74db97766aa

which implements standard-conforming behavior according to

https://github.com/json-schema-org/JSON-Schema-Test-Suite/commit/b24fae28923ba15e9199fc42c6976b2dd4125b46

Fix: explicitly combine requirements with AllOf instead of relying on
non-conforming "merging" behavior.

This change also adds a host test for cmc so it'll run on CQ.

CF-595 #done

Change-Id: If497820c794a7e2b4c16439a6e37fba6130aaaee