[dispatcher-pool] Fix more ASSERT preconditions.
Fix another violation of an invariant which was triggering an ASSERT
in EventSource::InternalDeactiveLocked.

Background:

EventSources exist in 1 of 4 states.  They are..
++ Idle: The source is not currently being waited on, or dispached.
++ WaitingOnPort: The handle to the event source has been sent to the
   port with wait_async, and is either in the kernel, or on its way up
   through the stack to the execution domain.
++ DispatchPending: The EventSource's handle was signaled, and the
   source has been placed in its execution domain's
   pending_dispatch_queue_
++ Dispatching: The EventSource has been removed from the pending
   queue by a thread assigned to the execution domain for the purposes
   of dispatching, and the dispatch is currently in progress.  The
   event source will return to either Idle or WaitingOnPort after
   this.

One of the invariants here is that EventSources only exist in their
execution domain's pending_dispatch_queue_ if and only if their state
is DispatchPending.  During deactivation of an event source, there is
an ASSERT to this effect.  The EventSource is unconditionally canceled
out of the pending queue, then if the state of the EventSource is not
WaitingOnPort, the code asserts that its state must be either Idle or
Dispatching.  Now that it is no longer in the pending queue, its state
_cannot_ be DispatchPending.

The Bug:

This invariant was being violated and caught.  One possible flow which
shows this problem is as follows.

1) An event source for an execution domain has fired and is in the
   process of executing on thread A.
2) Another event source fires and gets added to the pending queue by
   thread B.
3) Thread C (a thread from outside of the thread pool) calls
   Deactivate on the ExecutionDomain that the two event sources belong
   to.
4) Thread C enters the source lock for the execution domain, sets the
   deactivated flag, and swaps the domain's source list into a local
   list.  Now that the domain is deactivated, no new sources can join
   the domain.  It then leaves the source lock.
5) Thread C is about to iterate over the sources which had been in the
   domain and deactivate each one.  Before it does, however...
6) Thread A returns from its dispatch of the first event source.  It
   enters the source lock in order to find the next source to
   dispatch.  (this is in ExecutionDomain::Dispatch)
7) Thread A discovers that the execution domain has been deactivated.
   Instead of dispatching the next pending event source, it fast
   aborts.
8) During the fast abort, the it signals that it is done dispatching,
   which is needed for synchronizing with thread C.  In the process,
   however, it also clears the pending_work_queue_.
9) At this point, the invariant has been violated.  Any EventSources
   which were in the pending event queue now are not anymore, but
   their state is still DispatchPending.
10) Back in Thread C land, thread C attempts to deactivate the next
    event source.  It calls CancelPendingLocked.
11) CancelPendingLocked attempts to remove the EventSource from the
    pending queue, and discovers that it is not in the pending queue.
    The state of the EventSource is not WaitingOnPort either, so it
    signals that it is OK for the caller to close the handle now.
12) Just before Thread C closes the handle, this ASSERT fires.  After
    CancelPendingLocked, the EventSource can be in any state but
    DispatchPending, but the bookkeeping says otherwise.

The Fix:

The fix is simple.  If a domain has become deactivated, and a thread
which is dispatching pending event sources for the domain is
unwinding, it simply should not clear the pending_work_queue_.  This
preserves the invariant, and the event sources are in the process of
being removed from the queue during deactivation anyway.

The Impact:

No real impact here.  The domain was being shutdown and all of the
event sources deactivated anyway.  There internal state is not really
much of an issue at this point in time.  Had this been a release
build, everything would have been cleaned up properly, even when this
edge case triggered.  Technically, the invariant could be declared not
useful, and the ASSERT could have been viewed as overly aggressive and
removed.  Instead, I think it is better to preserve the invariant,
since it actually results in doing less work, and allows us to keep
the assert as a safety.

Testing:
Repro'ed The bug using the steps in ZX-3424 in a while true shell
loop.  Usually took about 3-4 hours (I caught 3 in a single overnight
run).  I then instrumented the assert to stall instead of halt,
allowing me to attach with ZXDB and work out the cause by examining
the state of the system interactively.  After finding the root cause
via inspection, the patch was written and again tested overnight.
The test loop has been running for 17.9hrs now with no failures.

Tests: See "Testing" above.
Change-Id: I4321a420facc757a4f96725607fe8ac0e3fee661